+++ backend/src/services/team.service.ts
@@ -3,12 +3,13 @@ import { RequestError } from "../@types/requestError.js";
 import { prisma } from "../config/prisma.js";
 import logger from "../config/logger.js";
 import { CACHE_KEY_PREFIXES, generateCacheKey, redisClient } from "../config/redis.js";
-import { BigIntreplacer, invalidateCache, isValidTeamId, updateCacheData } from "../utils/validate.functions.js";
+import { BigIntreplacer, isValidTeamId } from "../utils/validate.functions.js";
+import { invalidateCache, updateCacheData } from "../config/redis.js"
 import { setJoinedTeamsTypeBody, setTeamsTypeBody } from "../schemas/team.schema.js";
 import fs from "fs/promises";
 import path from "path";
 import { FINAL_UPLOADS_DIR } from "../config/upload.js";
-import socketIO, { SOCKET_EVENTS } from "../config/socket.js";
+import { emitToClass, SOCKET_EVENTS } from "../config/socket.js";
 
 const teamService = {
   async getTeamsData(session: Session & Partial<SessionData>) {
@@ -174,25 +175,24 @@ const teamService = {
     }
 
     if (dataChanged) {
-      // invalidate team cache
-      await invalidateCache("TEAMS", classId.toString());
-      const io = socketIO.getIO();
-      io.to(`class:${session.classId}`).emit(SOCKET_EVENTS.TEAMS);
-      io.to(`class:${session.classId}`).emit(SOCKET_EVENTS.JOINED_TEAMS);
+      // invalidate team cache and resend sockets
+      await invalidateCache(CACHE_KEY_PREFIXES.TEAMS, classId.toString());
+      emitToClass(classId, SOCKET_EVENTS.TEAMS);
+      emitToClass(classId, SOCKET_EVENTS.JOINED_TEAMS);
 
-      // If teams were deleted, also update homework, events, lesson and upload (request) caches
+      // If teams were deleted, also update homework, events, lesson and upload (request) caches and resend sockets
       if (teamsDeleted) {
-        await invalidateCache("HOMEWORK", session.classId!);
-        await invalidateCache("EVENT", session.classId!);
-        await invalidateCache("LESSON", session.classId!);
-        await invalidateCache("UPLOADMETADATA", session.classId!);
-        await invalidateCache("UPLOADREQUESTS", session.classId!);
+        await invalidateCache(CACHE_KEY_PREFIXES.HOMEWORK, classId.toString());
+        await invalidateCache(CACHE_KEY_PREFIXES.EVENT, classId.toString());
+        await invalidateCache(CACHE_KEY_PREFIXES.LESSON, classId.toString());
+        await invalidateCache(CACHE_KEY_PREFIXES.UPLOADMETADATA, classId.toString());
+        await invalidateCache(CACHE_KEY_PREFIXES.UPLOADREQUESTS, classId.toString());
 
-        io.to(`class:${session.classId}`).emit(SOCKET_EVENTS.HOMEWORK);
-        io.to(`class:${session.classId}`).emit(SOCKET_EVENTS.EVENTS);
-        io.to(`class:${session.classId}`).emit(SOCKET_EVENTS.TIMETABLES);
-        io.to(`class:${session.classId}`).emit(SOCKET_EVENTS.UPLOADS);
-        io.to(`class:${session.classId}`).emit(SOCKET_EVENTS.UPLOAD_REQUESTS);
+        emitToClass(classId, SOCKET_EVENTS.HOMEWORK);
+        emitToClass(classId, SOCKET_EVENTS.EVENTS);
+        emitToClass(classId, SOCKET_EVENTS.TIMETABLES);
+        emitToClass(classId, SOCKET_EVENTS.UPLOADS);
+        emitToClass(classId, SOCKET_EVENTS.UPLOAD_REQUESTS);
       }
       logger.info(`teams data changed for class: ${classId}`);
     }
@@ -217,6 +217,7 @@ const teamService = {
   async setJoinedTeamsData(reqData: setJoinedTeamsTypeBody, session: Session & Partial<SessionData>) {
     const { teams } = reqData;
     const accountId = session.account!.accountId;
+    const classId = parseInt(session.classId!, 10);
 
     await prisma.$transaction(async tx => {
       await tx.joinedTeams.deleteMany({
@@ -234,9 +235,8 @@ const teamService = {
         });
       }
     });
-
-    const io = socketIO.getIO();
-    io.to(`class:${session.classId}`).emit(SOCKET_EVENTS.JOINED_TEAMS);
+    // send socket update to clients
+    emitToClass(classId, SOCKET_EVENTS.JOINED_TEAMS);
   }
 };